Exception handling

+++ src/main/java/com/endava/wallet/controller/UserController.java
@@ -2,6 +2,7 @@ package com.endava.wallet.controller;
 
 import com.endava.wallet.entity.Authority;
 import com.endava.wallet.entity.User;
+import com.endava.wallet.exception.ApiRequestException;
 import com.endava.wallet.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -25,8 +26,11 @@ public class UserController {
         return "userList";
     }
 
-    @GetMapping("{user}")
-    public String userEditForm(@PathVariable User user, Model model) {
+    @GetMapping("{userID}")
+    public String userEditForm(@PathVariable Long userID, Model model) {
+        User user = userService.findUserById(userID);
+        if (user == null)
+            throw new ApiRequestException("There is no user with such id in database");
         model.addAttribute("user", user);
         model.addAttribute("authorities", Authority.values());
         return "userEdit";
@@ -36,14 +40,19 @@ public class UserController {
     public String userSave(
             @RequestParam String username,
             @RequestParam Map<String, String> form,
-            @RequestParam Long id
+            @RequestParam Long userID
     ) {
-        userService.add(username, form, userService.findUserById(id));
+        User user = userService.findUserById(userID);
+        if (user == null)
+            throw new ApiRequestException("There is no user with id = " + userID + " in database");
+        userService.add(username, form, user);
         return "redirect:/user";
     }
 
     @GetMapping("/delete/{userID}")
     public String userDelete(@PathVariable Long userID) {
+        if (userService.findUserById(userID) == null)
+            throw new ApiRequestException("There is no user with id = " + userID + " in database");
         userService.deleteUserById(userID);
         return "redirect:/user";
     }
